Refactor
Some Issues with wrong assumptions. The log shuffling scheme initially
did not work because the metronome struct was constantly nil and not
updated properly thus flushing immediately. As such the straggler
handling did not work as well as we were never flushing.

With flushing there was an issue with the log prefix changing due to
other goroutines changing it. Having a proper copy of the prefix fixed
the issue