Implement recovery
Very rough and not performed implementation of the algorithm.
Lacks some more performant data structs especially for the
in-mem entries. Also lacks consideration for only asking a majority
quorum for missing entries on recovery. Furthermore needs implementation
on reclaiming memory on snapshots etc so that the unflushed entries do
not grow indefinetely